OXDEV-351 Remove logging in handleRoutingException
If the debug mode was enabled in config.inc.php, the method \OxidEsales\EshopCommunity\Core\ShopControl::handleRoutingException() logged whenever a controller id could not be resolved to a class name. As long as the BC layer is still part of the OXID eShop, it is an expected behavior that controller ids could not be resolved and therefore the logging of this exception is excessive.